feat: add deps file
